requirements.txt
astroid==2.3.3
blinker==1.4
certifi==2019.11.28
chardet==3.0.4
Click==7.0
DateTime==4.3
dateutils==0.6.7
Flask==1.1.1
Flask-Mail==0.9.1
idna==2.8
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.3
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
passlib==1.7.2
PyJWT==1.7.1
pylint==2.4.4
python-dateutil==2.8.1
pytz==2019.3
requests==2.22.0
six==1.13.0
twilio==6.34.0
typed-ast==1.4.0
urllib3==1.25.7
Werkzeug==0.16.0
wrapt==1.11.2
zope.interface==4.7.1
